feat(runtime): int64 budget, versioned checkpoints, bounded event log
Three pre-calcification fixes applied before the codebase hardens:

1. float64 → int64 microcents for all budget/price fields (RE-3 compliance).
   New pkg/budget package provides FromFloat/Format helpers. Cost calculation
   uses pure integer arithmetic: elapsed_us × price / 1_000_000.

2. Checkpoint format v1: [version:1][budget:8][price:8][state:N] with
   version byte at offset 0. Unsupported versions rejected explicitly.

3. Bounded event log: NewEventLog(maxTicks) evicts oldest ticks when
   history exceeds limit. DefaultMaxTicks=1024 (~17 min at 1Hz).

Also includes Phase 2 hardening: race condition fix in migration service
(sync.RWMutex), panic recovery in tick loop (EI-6), path traversal
validation in storage, and new test suites for storage/runtime/agent.

+++ pkg/manifest/manifest.go
@@ -58,8 +58,8 @@ type MigrationPolicy struct {
 	// Enabled indicates whether migration is allowed.
 	Enabled bool
 
-	// MaxPricePerSecond is the maximum price the agent is willing to pay.
-	MaxPricePerSecond float64
+	// MaxPricePerSecond is the maximum price the agent is willing to pay (microcents).
+	MaxPricePerSecond int64
 
 	// PreferredRegions is a list of geographic preferences (optional).
 	PreferredRegions []string
